requirements.txt
streamlit==1.42.2
langchain==0.3.17
langchain-openai==0.3.5
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.3
python-dotenv==1.0.1
pyyaml==6.0.2
